+++ src/ArrowDbCore/ArrowDb.cs
@@ -97,6 +97,7 @@ public sealed partial class ArrowDb {
     /// <remarks>
     /// The <see cref="ArrowDbTransactionScope"/> implements both <see cref="IDisposable"/> and <see cref="IAsyncDisposable"/>, allowing it to be used in both synchronous and asynchronous contexts.
     /// </remarks>
+    /// <param name="cancellationToken">A cancellation token for the outermost implicit serialize operation.</param>
     /// <returns>A new <see cref="ArrowDbTransactionScope"/> instance.</returns>
-    public ArrowDbTransactionScope BeginTransaction() => new(this);
+    public ArrowDbTransactionScope BeginTransaction(CancellationToken cancellationToken = default) => new(this, cancellationToken);
 }