+++ src/ArrowDbCore/Serializers/InMemorySerializer.cs
@@ -10,10 +10,10 @@ public sealed class InMemorySerializer : IDbSerializer {
     /// <summary>
     /// Returns an empty dictionary
     /// </summary>
-    public ValueTask<ConcurrentDictionary<string, byte[]>> DeserializeAsync() => ValueTask.FromResult(new ConcurrentDictionary<string, byte[]>());
+    public ValueTask<ConcurrentDictionary<string, byte[]>> DeserializeAsync(CancellationToken cancellationToken = default) => ValueTask.FromResult(new ConcurrentDictionary<string, byte[]>());
 
     /// <summary>
     /// Does nothing
     /// </summary>
-    public ValueTask SerializeAsync(ConcurrentDictionary<string, byte[]> data) => ValueTask.CompletedTask;
+    public ValueTask SerializeAsync(ConcurrentDictionary<string, byte[]> data, CancellationToken cancellationToken = default) => ValueTask.CompletedTask;
 }